+++ test/metadata/1.0.0/types.test.js
@@ -1,5 +1,5 @@
 import metadata from './metadata.full'
-import get_number_type_custom from '../../../source/types'
+import get_number_type_custom from '../../../source/getNumberType'
 
 function get_number_type(...parameters)
 {